// web/src/api/generated/schemas/ratingPreviewResource.zod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Fonzytooter API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod'

export const ratingPreviewResourceIntervalDaysMin = 0

export const ratingPreviewResourceIntervalSecondsMin = 0

export const RatingPreviewResource = zod.strictObject({
  dueAt: zod.iso.datetime({ offset: true }),
  intervalDays: zod.number().min(ratingPreviewResourceIntervalDaysMin),
  intervalSeconds: zod.int().min(ratingPreviewResourceIntervalSecondsMin),
  rating: zod.enum(['again', 'hard', 'good', 'easy']),
})

export type RatingPreviewResource = zod.input<typeof RatingPreviewResource>
export type RatingPreviewResourceOutput = zod.output<typeof RatingPreviewResource>