// web/src/api/generated/schemas/reviewSubmission.zod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Helix Academy API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod'

export const ReviewSubmission = zod.strictObject({
  $schema: zod.url().optional().describe('A URL to the JSON Schema for this object.'),
  rating: zod.enum(['again', 'hard', 'good', 'easy']),
})

export type ReviewSubmission = zod.input<typeof ReviewSubmission>
export type ReviewSubmissionOutput = zod.output<typeof ReviewSubmission>